Second pass at spec-compliant Promise
The first attempt was just a bluthering rush into implementing a
Promise. It meant I had to figure out the API and all its edge
cases and rules, *and* implement it, *and* figure out tests.

This second pass is more to exercise the second part - implementing
the spec - instead of doing all three parts.